wrap handler classes into util.
