--- ksql-common/src/main/java/io/confluent/ksql/errors/LogAndFailProductionExceptionHandler.java
@@ -1,25 +1,0 @@
-/*
- * Copyright 2019 Confluent Inc.
- *
- * Licensed under the Confluent Community License; you may not use this file
- * except in compliance with the License.  You may obtain a copy of the License at
- *
- * http://www.confluent.io/confluent-community-license
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package io.confluent.ksql.errors;
-
-import io.confluent.ksql.errors.ProductionExceptionHandlerUtil.LogAndXProductionExceptionHandler;
-
-public class LogAndFailProductionExceptionHandler extends LogAndXProductionExceptionHandler {
-
-  @Override
-  ProductionExceptionHandlerResponse getResponse() {
-    return ProductionExceptionHandlerResponse.FAIL;
-  }
-}
